+++ stdlib/ballerina-builtin/src/main/ballerina/ballerina/time/timelib.bal
@@ -1,4 +1,4 @@
-package ballerina.builtin;
+package ballerina.time;
 
 @Description {value:"Ballerina Timezone struct represents the timezone information associated with a particular time."}
 @Field {value:"zoneId: Zone short ID or offset string."}
