+++ stdlib/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/Utils.java
@@ -37,7 +37,7 @@ import java.util.TimeZone;
  */
 public class Utils {
 
-    public static final String PACKAGE_BUILTIN = "ballerina.builtin";
+    public static final String PACKAGE_TIME = "ballerina.time";
     public static final String STRUCT_TYPE_TIME = "Time";
     public static final String STRUCT_TYPE_TIMEZONE = "Timezone";
 
@@ -63,12 +63,18 @@ public class Utils {
     }
 
     public static StructInfo getTimeZoneStructInfo(Context context) {
-        PackageInfo timePackageInfo = context.getProgramFile().getPackageInfo(PACKAGE_BUILTIN);
+        PackageInfo timePackageInfo = context.getProgramFile().getPackageInfo(PACKAGE_TIME);
+        if (timePackageInfo == null) {
+            return null;
+        }
         return timePackageInfo.getStructInfo(STRUCT_TYPE_TIMEZONE);
     }
 
     public static StructInfo getTimeStructInfo(Context context) {
-        PackageInfo timePackageInfo = context.getProgramFile().getPackageInfo(PACKAGE_BUILTIN);
+        PackageInfo timePackageInfo = context.getProgramFile().getPackageInfo(PACKAGE_TIME);
+        if (timePackageInfo == null) {
+            return null;
+        }
         return timePackageInfo.getStructInfo(STRUCT_TYPE_TIME);
     }
