+++ stdlib/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/time/AbstractTimeFunction.java
@@ -15,7 +15,7 @@
 * specific language governing permissions and limitations
 * under the License.
 */
-package org.ballerinalang.nativeimpl.builtin.timelib;
+package org.ballerinalang.nativeimpl.time;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.values.BStruct;
