+++ stdlib/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/time/CreateTime.java
@@ -15,7 +15,7 @@
 * specific language governing permissions and limitations
 * under the License.
 */
-package org.ballerinalang.nativeimpl.builtin.timelib;
+package org.ballerinalang.nativeimpl.time;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
@@ -30,7 +30,7 @@ import org.ballerinalang.natives.annotations.ReturnType;
  * @since 0.89
  */
 @BallerinaFunction(
-        packageName = "ballerina.builtin",
+        packageName = "ballerina.time",
         functionName = "createTime",
         args = {@Argument(name = "years", type = TypeKind.INT),
                 @Argument(name = "months", type = TypeKind.INT),
@@ -41,7 +41,7 @@ import org.ballerinalang.natives.annotations.ReturnType;
                 @Argument(name = "milliseconds", type = TypeKind.INT),
                 @Argument(name = "zoneID", type = TypeKind.STRING)},
         returnType = {@ReturnType(type = TypeKind.STRUCT, structType = "Time",
-                                  structPackage = "ballerina.builtin")},
+                                  structPackage = "ballerina.time")},
         isPublic = true
 )
 public class CreateTime extends AbstractTimeFunction {
