+++ stdlib/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/time/Hour.java
@@ -15,7 +15,7 @@
 * specific language governing permissions and limitations
 * under the License.
 */
-package org.ballerinalang.nativeimpl.builtin.timelib;
+package org.ballerinalang.nativeimpl.time;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
@@ -27,27 +27,23 @@ import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.ReturnType;
 
 /**
- * Get the hour, minute, second and millisecond value for the given time.
+ * Get the hour value for the given time.
  *
  * @since 0.89
  */
 @BallerinaFunction(
-        packageName = "ballerina.builtin",
-        functionName = "Time.getTime",
+        packageName = "ballerina.time",
+        functionName = "Time.hour",
         args = {@Argument(name = "time", type = TypeKind.STRUCT, structType = "Time",
-                          structPackage = "ballerina.builtin")},
-        returnType = {@ReturnType(type = TypeKind.INT),
-                      @ReturnType(type = TypeKind.INT),
-                      @ReturnType(type = TypeKind.INT),
-                      @ReturnType(type = TypeKind.INT)},
+                          structPackage = "ballerina.time")},
+        returnType = {@ReturnType(type = TypeKind.INT)},
         isPublic = true
 )
-public class GetTime extends AbstractTimeFunction {
+public class Hour extends AbstractTimeFunction {
 
     @Override
     public BValue[] execute(Context context) {
         BStruct timeStruct = ((BStruct) getRefArgument(context, 0));
-        return getBValues(new BInteger(getHour(timeStruct)), new BInteger(getMinute(timeStruct)),
-                new BInteger(getSecond(timeStruct)), new BInteger(getMilliSecond(timeStruct)));
+        return new BValue[]{new BInteger(getHour(timeStruct))};
     }
 }
