+++ stdlib/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/time/MilliSecond.java
@@ -15,11 +15,11 @@
 * specific language governing permissions and limitations
 * under the License.
 */
-package org.ballerinalang.nativeimpl.builtin.timelib;
+package org.ballerinalang.nativeimpl.time;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.model.values.BString;
+import org.ballerinalang.model.values.BInteger;
 import org.ballerinalang.model.values.BStruct;
 import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.natives.annotations.Argument;
@@ -27,23 +27,23 @@ import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.ReturnType;
 
 /**
- * Convert a Time to ISO 8601 formatted string.
+ * Get the milli second value for the given time.
  *
  * @since 0.89
  */
 @BallerinaFunction(
-        packageName = "ballerina.builtin",
-        functionName = "Time.toString",
+        packageName = "ballerina.time",
+        functionName = "Time.milliSecond",
         args = {@Argument(name = "time", type = TypeKind.STRUCT, structType = "Time",
-                          structPackage = "ballerina.builtin")},
-        returnType = {@ReturnType(type = TypeKind.STRING)},
+                          structPackage = "ballerina.time")},
+        returnType = {@ReturnType(type = TypeKind.INT)},
         isPublic = true
 )
-public class ToString extends AbstractTimeFunction {
+public class MilliSecond extends AbstractTimeFunction {
+
     @Override
     public BValue[] execute(Context context) {
-
         BStruct timeStruct = ((BStruct) getRefArgument(context, 0));
-        return new BValue[]{new BString(getDefaultString(timeStruct))};
+        return new BValue[]{new BInteger(getMilliSecond(timeStruct))};
     }
 }
