+++ stdlib/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/time/ToTimezone.java
@@ -15,11 +15,10 @@
 * specific language governing permissions and limitations
 * under the License.
 */
-package org.ballerinalang.nativeimpl.builtin.timelib;
+package org.ballerinalang.nativeimpl.time;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.model.values.BInteger;
 import org.ballerinalang.model.values.BStruct;
 import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.natives.annotations.Argument;
@@ -27,23 +26,26 @@ import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.ReturnType;
 
 /**
- * Get the hour value for the given time.
+ * Change the timezone associated with the given time.
  *
  * @since 0.89
  */
 @BallerinaFunction(
-        packageName = "ballerina.builtin",
-        functionName = "Time.hour",
+        packageName = "ballerina.time",
+        functionName = "Time.toTimezone",
         args = {@Argument(name = "time", type = TypeKind.STRUCT, structType = "Time",
-                          structPackage = "ballerina.builtin")},
-        returnType = {@ReturnType(type = TypeKind.INT)},
+                          structPackage = "ballerina.time"),
+                @Argument(name = "zoneId", type = TypeKind.STRING)},
+        returnType = {@ReturnType(type = TypeKind.STRUCT, structType = "Time",
+                                  structPackage = "ballerina.time")},
         isPublic = true
 )
-public class Hour extends AbstractTimeFunction {
+public class ToTimezone extends  AbstractTimeFunction {
 
     @Override
     public BValue[] execute(Context context) {
         BStruct timeStruct = ((BStruct) getRefArgument(context, 0));
-        return new BValue[]{new BInteger(getHour(timeStruct))};
+        String zoneId = getStringArgument(context, 0);
+        return new BValue[] { changeTimezone(context, timeStruct, zoneId) };
     }
 }
