+++ tests/ballerina-test/src/test/resources/test-src/expressions/lambda/function-pointer-native.bal
@@ -1,6 +1,8 @@
+import ballerina.time;
+
 function test1()(int){
-    function (string, string) returns (Time) parseFunction = parse;
-    Time c = parseFunction("2017-07-20T00:00:00.000-0500", "yyyy-MM-dd'T'HH:mm:ss.SSSZ");
+    function (string, string) returns (time:Time) parseFunction = time:parse;
+    time:Time c = parseFunction("2017-07-20T00:00:00.000-0500", "yyyy-MM-dd'T'HH:mm:ss.SSSZ");
     return c.time;
 }
