+++ tests/ballerina-test/src/test/resources/test-src/natives/builtin-symbol-negative.bal
@@ -3,30 +3,13 @@ function test0 (){
 }
 
 function error (string x, error e){
-    print(x);
+    e.message = x;
 }
-
-function println(any a){
-    print(a);
-    print("\n");
-}
-
-
 
 function test1(){
-    int Time = 10;
+    int error = 10;
 }
 
 function <error e> getMessage() returns (string){
     return e.message;
-}
-
-function test2(){
-    function (any a) foo = print;
-    foo("abc");
-}
-
-enum Time {
-    VALUE1,
-    VALUE2
 }
